ci: fix ci for Node v14 by updating npm
NPM versions before 7 do not automatically install `"peerDependencies"`
when running `npm install` locally, so do not work with this repo
(at least unless we copy the peerDependencies to our devDependencies).

Because the GitHub `actions/setup-node@v3` comes with NPM v6 for
Node v14, this means we need to manually upgrade to NPM@>=7 for
CI to run.